convert to device
