+++ resources/web/wwi/nodes/WbCoordinate.js
@@ -1,7 +1,6 @@
 import WbBaseNode from './WbBaseNode.js';
 import WbWorld from './WbWorld.js';
 
-// This class is used to retrieve the type of device
 export default class WbCoordinate extends WbBaseNode {
   #point;
   constructor(id, point) {
