+++ resources/web/wwi/nodes/WbGyro.js
@@ -1,5 +1,5 @@
-import WbSolid from './WbSolid.js';
+import WbDevice from './WbDevice.js';
 
 // This class is used to retrieve the type of device
-export default class WbGyro extends WbSolid {
+export default class WbGyro extends WbDevice {
 }
