+++ resources/web/wwi/nodes/WbLed.js
@@ -1,5 +1,5 @@
-import WbSolid from './WbSolid.js';
+import WbDevice from './WbDevice.js';
 
 // This class is used to retrieve the type of device
-export default class WbLed extends WbSolid {
+export default class WbLed extends WbDevice {
 }
