+++ resources/web/wwi/nodes/WbReceiver.js
@@ -1,5 +1,5 @@
-import WbSolid from './WbSolid.js';
+import WbDevice from './WbDevice.js';
 
 // This class is used to retrieve the type of device
-export default class WbReceiver extends WbSolid {
+export default class WbReceiver extends WbDevice {
 }
